requirements.txt
Django==1.6.5
djangorestframework==2.3.13
Jinja2==2.7.2
MarkupSafe==0.21
Pillow==2.4.0
Wand==0.3.7
Werkzeug==0.9.4
aniso8601==0.82
itsdangerous==0.24
pilbox==0.9.18
pymongo==2.7
pytz==2014.2
six==1.6.1
sphinx-me==0.2.1
tornado==3.1.1
pbr==0.8.0
python-keystoneclient==0.8.0
python-swiftclient==2.0.3
requests==2.3.0
sentinels==0.0.6
simplejson==3.5.0
oslo.config==1.3.0
netaddr==0.7.11
py==1.4.20
prettytable==0.7.2
Babel==1.3
